Conditional request if it is or not favorite

+++ client/src/clientComponents/LikeButton/LikeButton.jsx
@@ -1,12 +1,16 @@
 import styles from "./LikeButton.module.css";
 import { useEffect } from "react";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { useAuth0 } from "@auth0/auth0-react";
 import { getUserFavoritesAction } from "../../redux/userSlice";
 import axios from "axios";
 
 export default function LikeButton({ foodId }) {
   const dispatch = useDispatch();
+  const favorites = useSelector((state) => state.usersReducer.userFavorites);
+
+  const isFavorite = favorites.some(favorite => favorite.id === foodId);
+
   const { user, isAuthenticated } = useAuth0();
 
   const handleLike = async () => {
@@ -16,12 +20,13 @@ export default function LikeButton({ foodId }) {
     }
     const email = user.email;
 
-    
-    // await axios.post("/favorite", { email, foodId });
-    // dispatch(getUserFavoritesAction(email));
-
-    await axios.delete(`/favorite/${email}/${foodId}`);
-    dispatch(getUserFavoritesAction(email));
+    if (isFavorite) {
+      await axios.delete(`/favorite/${email}/${foodId}`);
+      dispatch(getUserFavoritesAction(email));
+    } else {
+      await axios.post("/favorite", { email, foodId });
+      dispatch(getUserFavoritesAction(email));
+    }
   };
 
   return (
